% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_intervals.R
\name{pred_intervals}
\alias{pred_intervals}
\title{Prediction intervals for bart-bma output}
\usage{
pred_intervals(object, num_iter, burnin, l_quant, u_quant,
  newdata = NULL)
}
\arguments{
\item{object}{bartBMA object obtained from function bartBMA}

\item{num_iter}{Total number of iterations of the Gibbs sampler (including burn-in).}

\item{burnin}{Number of burn-on iterations of the Gibbs sampler.}

\item{l_quant}{Lower quartile of the prediction interval.}

\item{u_quant}{Upper quartile of the prediction interval.}

\item{newdata}{Test data for which predictions are to be prodiced. Default = NULL. If NULL, then produces prediction intervals for training data if no test data was used in producing the bartBMA object, or produces prediction intervals for the original test data if test data was used in producing the bartBMA object.}
}
\value{
The output is a list of length one. The one element in this list is a vector of prediction intervals???
}
\description{
This function produces prediction intervals for BART-BMA estimates by post-hoc Gibbs-sampling from the full conditionals of the terminal node parameters and the variance of the error term. See Hernandez et al. (2018) Appendix D for details.
}
