% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bartBMA.R
\name{predict.bartBMA}
\alias{predict.bartBMA}
\title{Predictions for a new dataset using an existing bartbma object}
\usage{
\method{predict}{bartBMA}(object, newdata)
}
\arguments{
\item{object}{A bartBMA object obtained using the barBMA function.}

\item{newdata}{Covariate matrix for new dataset.}
}
\value{
A vector of predictions for the new dataset.
}
\description{
This function produces predictions for a new dataset using a previously obtained bartBMA object.
}
